Add a script to revert Ubuntu to minimal installation
Reverts an Ubuntu installation on my old MacBook Pro from a standard to a minimal setup. Initially installed with the full suite of applications and utilities, but later figured out I won't be using them.

The script removes both explicitly visible and background applications that are not part of Ubuntu's minimal installation criteria.